requirements.txt
Flask==3.0.0
psycopg2==2.9.9
boto3==1.34.11
passlib==1.7.4
PyJWT
python-dotenv
flask-cors